% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nouns_api.R
\name{check_np_response}
\alias{check_np_response}
\title{Check Noun Project Response}
\usage{
check_np_response(resp)
}
\arguments{
\item{resp}{response from np api}
}
\description{
Check Noun Project Response
}
\examples{
\dontrun{No example needed}
}
