% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{display_icon}
\alias{display_icon}
\title{Display icon}
\usage{
display_icon(icon_number)
}
\arguments{
\item{icon_number}{character or string with icon code eg. 13094}
}
\value{
image details and shows image in viewer
}
\description{
Display icon
}
\examples{
\dontrun{
display_icon(13094)
}
}
