% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{display_many_icons}
\alias{display_many_icons}
\title{Display many icons}
\usage{
display_many_icons(icon_name, num_of_imgs = 4)
}
\arguments{
\item{icon_name}{character with name of icon to search in nounproject}

\item{num_of_imgs}{number of images to display (default 4)}
}
\value{
group of images in a PNG format
}
\description{
Displays first num_of_imgs (default 4) icons from The Noun Project.
}
