% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{get_icon_by_term}
\alias{get_icon_by_term}
\title{Search Noun Project using a Term using Python authentication}
\usage{
get_icon_by_term(term, num_of_imgs = 4, limit_to_public_domain = 1,
  offset = 0)
}
\arguments{
\item{term}{word for which you would like to search on Noun Project site}

\item{num_of_imgs}{the number of images you would like to get (default = 4 )}

\item{limit_to_public_domain}{enter 0 for NO or 1 for YES (default YES)}

\item{offset}{how many images would you like to skip (default 0)}
}
\value{
gives a list of image details including urls from which they can be
accessed
}
\description{
Search Noun Project using a Term using Python authentication
}
\examples{
\dontrun{
get_icon_by_term("dog")
}
}
