% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_oauth_python.R
\name{get_icon_by_term_python}
\alias{get_icon_by_term_python}
\title{Search Noun Project using a Term using Python authentication}
\usage{
get_icon_by_term_python(key, secret, term, num_of_imgs = 2,
  limit_to_public_domain = 1, offset = 0)
}
\arguments{
\item{key}{provided when you register an app on Noun Project website}

\item{secret}{provided when you register an app on Noun Project website}

\item{term}{word for which you would like to search on Noun Project site}

\item{num_of_imgs}{the number of images you would like to get}

\item{limit_to_public_domain}{enter 0 for NO or 1 for YES}

\item{offset}{how many images would you like to skip}
}
\value{
gives a list of image details including urls from which they can be
accessed
}
\description{
Search Noun Project using a Term using Python authentication
}
\examples{
\dontrun{
key <- my_key
secret <- my_secret
term = "elephant"
icon_lists <- get_icon_by_term(term)  # gets details of two icons
elephants <- get_pngs_and_show(icon_lists)  # download icons
magick::image_append(elephants)  # show icons - two elephants
}
}
