% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icons.R
\name{get_icons_urls}
\alias{get_icons_urls}
\title{Get icons urls}
\usage{
get_icons_urls(icon_list)
}
\arguments{
\item{icon_list}{list of icons from \code{get_icon_by_term}}
}
\value{
list with field .$url containing url and
.$id with icon id as character
}
\description{
Get icons urls
}
