% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nouns_api.R
\name{get_nouns_api}
\alias{get_nouns_api}
\title{Access Noun Project API with authorisation}
\usage{
get_nouns_api(endpoint, baseurl = "http://api.thenounproject.com/")
}
\arguments{
\item{endpoint}{added information for specific urls on Noun Project API}

\item{baseurl}{url for the Noun Project API}
}
\value{
JSON object
}
\description{
Access Noun Project API with authorisation
}
\examples{

}
