% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_icons.R
\name{get_png}
\alias{get_png}
\title{Get a PNG file after Noun Project access}
\usage{
get_png(res_from_api)
}
\arguments{
\item{res_from_api}{}
}
\value{
a PNG pointer that can be shown in the viewer
}
\description{
Get a PNG file after Noun Project access
}
\examples{
# this will show an airplane
res <- get_nouns_api("icon/609")
get_png(res)
}
