% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_oauth_python.R
\name{make_icon_num_url}
\alias{make_icon_num_url}
\title{Make simple authorisation url for one icon number}
\usage{
make_icon_num_url(icon_num)
}
\arguments{
\item{icon_num}{a number of the icon you want to download}
}
\value{
the url assembled
}
\description{
Make simple authorisation url for one icon number
}
