% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_endpoints.R
\name{make_id_endpoint}
\alias{make_id_endpoint}
\title{Make the endpoint for an icon url by id}
\usage{
make_id_endpoint(id)
}
\arguments{
\item{id}{numeric id of icon}
}
\value{
partial url string
}
\description{
Accepts id as argument to return partial url
}
\details{
pastes
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
