% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_endpoints.R
\name{make_term_endpoint}
\alias{make_term_endpoint}
\title{Making the endpoint for searching Noun Project by Term}
\usage{
make_term_endpoint(term, limit_to_public_domain = 1, num_of_imgs = 2,
  offset = 0)
}
\arguments{
\item{term}{}

\item{limit_to_public_domain}{}

\item{num_of_imgs}{}

\item{offset}{}
}
\value{
an url to use with get_nouns_api()
}
\description{
Making the endpoint for searching Noun Project by Term
}
\examples{
# find four icons of dogs
url <- make_term_endpoint("dog", num_of_imgs = 4)
res <- get_nouns_api(url)
result <- httr::content(res)
get_pngs_and_show(result)
}
