% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_oauth.R
\name{np_credentials}
\alias{np_credentials}
\title{Set credentials as system variables}
\usage{
np_credentials(key, secret)
}
\arguments{
\item{key}{your key from https://thenounproject.com/developers/apps/}

\item{secret}{your secret from https://thenounproject.com/developers/apps/}
}
\value{
invisibly as System Environemtnal variables: `NOUNS_API_KEY` and `NOUNS_API_SECRET`
}
\description{
Set credentials as system variables
}
