% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_oauth.R
\name{np_oauth}
\alias{np_oauth}
\title{Title}
\usage{
np_oauth(key, secret, icon_num)
}
\arguments{
\item{key}{provided when you register an app on Noun Project website}

\item{secret}{provided when you register an app on Noun Project website}

\item{icon_num}{number on the Noun Project website that corresponds to an icon that 
you would like to download}
}
\value{
JSON object with links to icons
}
\description{
Title
}
