% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_oauth.R
\name{get_nouns_api}
\alias{get_nouns_api}
\title{Access Noun Project API with authorisation}
\usage{
get_nouns_api(endpoint, baseurl = "http://api.thenounproject.com/",
  app = nouns_app)
}
\arguments{
\item{endpoint}{}

\item{baseurl}{url for the Noun Project API}

\item{app}{httr app that supplies the authorisation including key and
signature
#'}
}
\value{
JSON object
}
\description{
Access Noun Project API with authorisation
}
\examples{



}
