% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_oauth.R
\name{get_pngs_and_show}
\alias{get_pngs_and_show}
\title{Download PNG for the icons in a list from get_icon_by_term()}
\usage{
get_pngs_and_show(icon_lists)
}
\arguments{
\item{icon_lists}{List of details about icons from Noun Project supplied by
the function get_icon_by_term() (may be useful for other functions too)}
}
\value{
group of images in a PNG format
}
\description{
Download PNG for the icons in a list from get_icon_by_term()
}
\examples{
elephants <- get_pngs_and_show(icon_lists)  # download icons
magick::image_append(elephants)  # show icons - two elephants
}
