% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_endpoints.R
\name{make_term_endpoint}
\alias{make_term_endpoint}
\title{Making the endpoint for searching Noun Project by Term}
\usage{
make_term_endpoint(term, limit_to_public_domain = 1, num_of_imgs = 2,
  offset = 0)
}
\arguments{
\item{term}{word to use for search}

\item{limit_to_public_domain}{restrict licence, change to 0 for not limit}

\item{num_of_imgs}{the number of images required}

\item{offset}{how many images to ignore, if you want to get different ones}
}
\value{
an url to use with get_nouns_api()
}
\description{
Making the endpoint for searching Noun Project by Term
}
\examples{
\dontrun{
# assemble the url to get 4 dog icons
url <- nounprojectR::make_term_endpoint("dog", num_of_imgs = 4)
# obtain information from the Noun Project server 
res <- get_nouns_api(url)
# extract the JSON file
result <- httr::content(res)
# show the images
get_pngs_and_show(result)
}
}
