% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_credentials.R
\name{np_credentials}
\alias{np_credentials}
\title{Set credentials as system variables}
\usage{
np_credentials(appname, key, secret)
}
\arguments{
\item{appname}{you appname from https://thenounproject.com/developers/apps/}

\item{key}{your key from https://thenounproject.com/developers/apps/}

\item{secret}{your secret from https://thenounproject.com/developers/apps/}
}
\value{
invisibly as System Environmental variables: `NOUNS_API_KEY` and `NOUNS_API_SECRET`
}
\description{
Set credentials as system variables
}
\examples{
np_credentials("appname", "your-noun-project-key", "your-noun-project-secret")
}
