% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_oauth.R
\name{np_oauth}
\alias{np_oauth}
\title{Get an icon by number using Python authentication}
\usage{
np_oauth(key, secret, url)
}
\arguments{
\item{key}{provided when you register an app on Noun Project website}

\item{secret}{provided when you register an app on Noun Project website}

\item{icon_num}{number on the Noun Project website that corresponds to an icon that 
you would like to download}
}
\value{
list (JSON object) with links to icons
}
\description{
Get an icon by number using Python authentication
}
\examples{
key <- my_key
secret <- my_secret
icon_number <- 1
output <- np_oauth(key, secret, icon_number) # output is a list
# access PNG using image_read() from magick
img2 <- magick::image_read(output$icon$preview_url)
img2  # show image - for icon_number of 1 it is a man putting litter in a bin



}
