% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/np_oauth_python.R
\name{np_oauth}
\alias{np_oauth}
\title{Get an icon by number using Python authentication}
\usage{
np_oauth(key, secret, url, python_path = Sys.which("python"))
}
\arguments{
\item{key}{provided when you register an app on Noun Project website}

\item{secret}{provided when you register an app on Noun Project website}

\item{url}{url created by make_icon_num_url() function.}

\item{python_path}{location of python on your computer}
}
\value{
list (JSON object) with links to icons
}
\description{
Get an icon by number using Python authentication
}
\examples{
\dontrun{
key <- my_key
secret <- my_secret
icon_number <- 1
url <- make_icon_num_url(icon_number)
output <- np_oauth(key, secret, url) # output is a list
# access PNG using image_read() from magick
img2 <- magick::image_read(output$icon$preview_url)
img2  # show image - for icon_number of 1 it is a man putting litter in a bin
}
}
